Feat: 수정 통계 감각별 퍼센트 실수형으로 변경 StatResponseDto.percentageOfCategory

+++ src/main/java/fivesenses/server/fivesenses/dto/StatResponseDto.java
@@ -18,7 +18,7 @@ import java.util.Map;
 public class StatResponseDto {
 
     private int totalPost;
-    private Map<Category, Integer> percentageOfCategory = new HashMap<>();
+    private Map<Category, Double> percentageOfCategory = new HashMap<>();
     private Map<Category, Long> cntOfCategory = new HashMap<>();
 
     private List<MonthlyMostCategoryDto> monthlyCategoryDtoList = new ArrayList<>();
